Don't crash on volume change
If changedEventCallback isn't defined and the user changes the volume, it will trigger an event that will raise an error since it's trying to call "sendPluginResult" on a null reference. This would happen if the app hasn't yet called "window.plugin.volume.getVolume" and a volume event is triggered by the system.